Move handling for init command into separate file.
